Update README.md
Readme:
A README file is typically a text file that provides information about a project. It is usually placed at the root of a project’s directory and is often the first place developers or users will look to understand the project’s purpose, how to use it, and any other important details.
Git:
Git is a distributed version control system (VCS) used to track changes in source code during software development. It allows multiple developers to work on a project simultaneously without interfering with each other’s work and provides a history of changes made to a project. Git is widely used in software development due to its powerful features, flexibility, and speed.
Github:
GitHub is a web-based platform that provides version control for software development and facilitates collaboration. It is built around Git, the distributed version control system (VCS), and enhances it with a range of additional features aimed at improving the development process, including project management, issue tracking, code review, and documentation.

GitHub is widely used by developers, open-source projects, and businesses for hosting and sharing code. It provides a cloud-based repository service, which allows users to store their Git repositories remotely and collaborate on them.